Expand rustc lints (#42)
Adds a number of additional lints as a `#![warn(...)]` directive,
including `trivial_numeric_casts`, which caught a number of
redundancies.

+++ src/uint/div.rs
@@ -19,7 +19,7 @@ impl<const LIMBS: usize> UInt<LIMBS> {
     /// When used with a fixed `rhs`, this function is constant-time with respect
     /// to `self`.
     pub(crate) const fn ct_div_rem(&self, rhs: &Self) -> (Self, Self, u8) {
-        let mut bd = self.bits().saturating_sub(rhs.bits()) as usize;
+        let mut bd = self.bits().saturating_sub(rhs.bits());
         let mut rem = *self;
         let mut quo = Self::ZERO;
 
@@ -59,7 +59,7 @@ impl<const LIMBS: usize> UInt<LIMBS> {
     /// When used with a fixed `rhs`, this function is constant-time with respect
     /// to `self`.
     pub(crate) const fn ct_reduce(&self, rhs: &Self) -> (Self, u8) {
-        let mut bd = self.bits().saturating_sub(rhs.bits()) as usize;
+        let mut bd = self.bits().saturating_sub(rhs.bits());
         let mut rem = *self;
 
         let mut c = rhs.shl_vartime(bd);
